Single Login for both operator and diner
