+++ auth/diner-auth-router.js
@@ -2,13 +2,10 @@ const router = require('express').Router();
 
 const bcrypt = require('bcryptjs');
 
-const jwt = require('jsonwebtoken');
-
 const Users = require('../diners/diners-model');
 
-const secrets = require('../config/secrets.js');
 
-router.post('/register', validateUserInfo, checkForUsername, checkPassword, (req, res) => {
+router.post('/register', validateUserInfo, checkForUsername, checkPassword, checkEmail, (req, res) => {
     const user = req.body;
 
     const hash = bcrypt.hashSync(user.password, 10);
@@ -26,42 +23,10 @@ router.post('/register', validateUserInfo, checkForUsername, checkPassword, (req
         
 });
 
-router.post('/login', validateLogin, (req,res) => {
-    let { username, password} = req.body;
-
-    Users.findBy({username})
-        .first()
-        .then(user => {
-            console.log(user)
-            if (user && bcrypt.compareSync(password, user.password)) {
-                const token = genToken(user);
-                res.status(200).json({username: user.username, department: user.department, token: token})
-            } else {
-                res.status(401).json({message: "Invalid credentials"})
-            }
-        })
-        .catch( err=> {
-            res.status(500).json(err)
-        })
-})
 
 
+/// Check for Username
 
-/// Helper function
-
-function genToken(user) {
-    const payload = {
-        userid: user.id,
-        username: user.username
-    }
-
-    const options = { expiresIn: '1h' };
-    const token = jwt.sign(payload, secrets.jwtSecret, options);
-
-  return token;
-}
-
-/// Validate User
 
 function checkForUsername(req, res, next) {
     let { username } = req.body;
@@ -74,9 +39,11 @@ function checkForUsername(req, res, next) {
                 next();
             }
     })    
-  }
+}
 
-  function validateUserInfo(req, res, next) {
+/// Validate UserInfo
+
+function validateUserInfo(req, res, next) {
     const postData = req.body;
     if(postData.username === "") {
         res.status(400).json({ message: "username can not be empty" });
@@ -89,18 +56,35 @@ function checkForUsername(req, res, next) {
     } else {
         next();
     }
-  }
+}
+
+/// Check Secure Password
 
 function checkPassword(req, res, next) { 
 
     let { password } = req.body;
-    var secureCheck = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[^a-zA-Z0-9])(?!.*\s).{8,50}$/;
+    const secureCheck = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[^a-zA-Z0-9])(?!.*\s).{8,50}$/;
         if(password.match(secureCheck)) { 
             next()
         } else { 
             res.status(400).json({ message: 'please enter a correct password'})
         }
 } 
+
+/// Check Valid Email
+
+function checkEmail(req, res, next) {
+    let {email} = req.body;
+    const emailFormat = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+
+    if(email.match(emailFormat)) {
+        next()
+    } else {
+        res.status(400).json({ message: 'please enter a correct email address'})
+    }
+}
+
+/// Check Valid Login Fields
 
 function validateLogin(req, res, next) {
     const postData = req.body;
